- creates `trapFocus` action based on `focus-trap` module - moves `dato-request` in `utils` folder
