+++ package.json
@@ -18,6 +18,7 @@
     "@lukeed/uuid": "^2.0.0",
     "cookie": "^0.4.1",
     "datocms-structured-text-to-dom-nodes": "^1.2.0",
+    "focus-trap": "^6.7.1",
     "stream": "0.0.2"
   }
 }
